+++ TA_BPMN_editor/app/index.js
@@ -9,7 +9,6 @@ import { debounce } from 'min-dash';
 
 import diagramXML from '../resources/newDiagram.bpmn';
 
-
 const fileSelector = document.getElementById('inputFile_bpmn');
 fileSelector.addEventListener('change', (event) => {
   const fileList = event.target.files;
@@ -23,27 +22,8 @@ fileSelector.addEventListener('change', (event) => {
       openDiagram(xml);
     };
     reader.readAsText(file);
-    $('#js-download-diagram').removeClass('active');
-    $('#js-download-svg').removeClass('active');
-    $('#js-download-intertasks').removeClass('active');
-  }
-  else {
-    window.alert("Format not valid: " + ext);
-  }
-});
 
-const customSelector = document.getElementById('inputFile_intertasks');
-customSelector.addEventListener('change', (event) => {
-  const fileList = event.target.files;
-  var file = fileList[0];
-  var ext = file.name.split('.').pop().toLowerCase();
-  if (ext === 'json') {
-    var reader = new FileReader();
-    reader.onload = function (e) {
-      var customElements = e.target.result;
-      bpmnModeler.addCustomElements(JSON.parse(customElements));
-    };
-    reader.readAsText(file);
+    fileSelector.disabled = true;
   }
   else {
     window.alert("Format not valid: " + ext);
@@ -51,10 +31,14 @@ customSelector.addEventListener('change', (event) => {
 });
 
 
+let commandStack_changed = false;
+let container = $('#js-drop-zone');
+document.getElementById('button-download-diagram').disabled = true;
+document.getElementById('button-download-svg').disabled = true;
 
-var container = $('#js-drop-zone');
 
-var bpmnModeler = new CustomModeler({
+
+let bpmnModeler = new CustomModeler({
   container: '#canvas',
   propertiesPanel: {
     parent: '#js-properties-panel'
@@ -72,23 +56,30 @@ var bpmnModeler = new CustomModeler({
   }
 });
 
+
 function createNewDiagram() {
   openDiagram(diagramXML);
 }
 
-
-async function openDiagram(xml, cElements) {
+async function openDiagram(xml) {
   try {
+
+
     await bpmnModeler.importXML(xml);
     bpmnModeler.cleanCustomElements();
-
-    if (cElements) {
-      bpmnModeler.addCustomElements(cElements);
-    }
+    bpmnModeler.loadCustomElementsFromXML();
 
     container
       .removeClass('with-error')
       .addClass('with-diagram');
+
+    bpmnModeler.setTCEvaluationsModulesButtons();
+
+    document.getElementById('button-download-diagram').disabled = false;
+    document.getElementById('button-download-svg').disabled = false;
+
+
+
   } catch (err) {
 
     container
@@ -151,11 +142,42 @@ $(function () {
     createNewDiagram();
   });
 
-  var downloadLink = $('#js-download-diagram');
-  var downloadSvgLink = $('#js-download-svg');
-  var downloadIntertasksLink = $('#js-download-intertasks');
 
-  bpmnModeler.setTCEvaluationsModulesButtons();
+  let buttonDownloadDiagram = $('#button-download-diagram');
+  let buttonDownloadSVG = $('#button-download-svg');
+
+
+  buttonDownloadDiagram.click(async function (e) {
+    try {
+      let definitions = bpmnModeler.getDefinitionsWithRelativeConstraintAsExtensionElements();
+      let { xml } = await bpmnModeler._moddle.toXML(definitions, { format: true });
+      downloadBPMN('diagram.bpmn', xml);
+
+    } catch (err) {
+      console.error('Error happened saving diagram: ', err);
+    }
+  });
+
+  buttonDownloadSVG.click(async function (e) {
+    try {
+      let { svg } = await bpmnModeler.saveSVG();
+      downloadBPMN('diagram.svg', svg);
+
+    } catch (err) {
+      console.error('Error happened saving SVG: ', err);
+    }
+  });
+
+  function downloadBPMN(filename, dataInput) {
+
+    let element = document.createElement('a');
+    // element.setAttribute('href','data:text/plain;charset=utf-8, ' + encodeURIComponent(dataInput));
+    element.setAttribute('href', 'data:application/bpmn20-xml;charset=UTF-8, ' + encodeURIComponent(dataInput));
+    element.setAttribute('download', filename);
+    document.body.appendChild(element);
+    element.click();
+    document.body.removeChild(element);
+  }
 
   $('.buttons a').click(function (e) {
     if (!$(this).is('.active')) {
@@ -164,49 +186,9 @@ $(function () {
     }
   });
 
-  function setEncoded(link, name, data) {
-    var encodedData = encodeURIComponent(data);
-
-    if (data) {
-      link.addClass('active').attr({
-        'href': 'data:application/bpmn20-xml;charset=UTF-8,' + encodedData,
-        'download': name
-      });
-    } else {
-      link.removeClass('active');
-    }
-  }
-
   var exportArtifacts = debounce(async function () {
 
-    try {
-      const { svg } = await bpmnModeler.saveSVG();
-      setEncoded(downloadSvgLink, 'diagram.svg', svg);
-
-    } catch (err) {
-      console.error('Error happened saving SVG: ', err);
-      setEncoded(downloadSvgLink, 'diagram.svg', null);
-    }
-
-    try {
-      let customElements = bpmnModeler.getCustomElements();
-      let myJson = JSON.stringify(customElements, null, '\t');
-      setEncoded(downloadIntertasksLink, 'intertasks.json', myJson);
-
-    } catch (err) {
-      console.error('Error happened saving intertasks: ', err);
-      setEncoded(downloadIntertasksLink, 'intertasks.json', null);
-    }
-
-    try {
-      let definitions = bpmnModeler.getDefinitions();
-      let { xml } = await bpmnModeler._moddle.toXML(definitions, { format: true });
-      setEncoded(downloadLink, 'diagram.bpmn', xml);
-
-    } catch (err) {
-      console.error('Error happened saving diagram: ', err);
-      setEncoded(downloadLink, 'diagram.bpmn', null);
-    }
+    commandStack_changed = true;
 
   }, 500);