+++ TA_BPMN_editor/app/temporal-modeler/modeler/CustomElementFactory.js
@@ -30,7 +30,7 @@ export default function CustomElementFactory(bpmnFactory, moddle) {
    */
   this.create = function(elementType, attrs) {
     var type = attrs.type;
-    // debugger;
+
     if (elementType === 'label') {
       return self.baseCreate(elementType, assign({ type: 'label' }, DEFAULT_LABEL_SIZE, attrs));
     }
@@ -51,7 +51,7 @@ export default function CustomElementFactory(bpmnFactory, moddle) {
           // https://forum.bpmn.io/t/customelements-autogenerate-id/3505
           
           assign(attrs.businessObject, {
-            id: 'InterTask_' + RENDERER_IDS.next()
+            id: 'InterTask_' + RENDERER_IDS.next().substring(0,4)
           });
         }
       }